backend/requirements.txt
flask
flask-cors
google-generativeai
pydantic
python-dotenv
groq
